requirements.txt
requests
feedparser
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth